trade handling bug fixed

+++ database/migrations/2022_06_12_120246_create_user_wallet_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->integer('user_id')->unsigned();
             $table->float('rial_balance');
-            $table->integer('mazin_balance')->unsigned();
+            $table->float('mazin_balance')->unsigned();
             $table->timestamps();
         });
     }
